refactor: remove in-memory advertisement store
