+++ src/structures/GrouperClient.js
@@ -1,8 +1,7 @@
 import { Client, ClientOptions } from 'discord.js';
 import TagStore from '../stores/TagStore';
-import AdvertisementStore from '../stores/AdvertisementStore';
 import CommandStore from '../stores/CommandStore';
-import GrouperCommandRouter from '../handlers/GrouperCommandRouter';
+import GrouperCommandRouter from './GrouperCommandRouter';
 
 export default class GrouperClient extends Client {
 
@@ -33,13 +32,6 @@ export default class GrouperClient extends Client {
          * @type {TagStore}
          */
         this.tags = new TagStore(this);
-
-        /**
-         * Client's advertisement store
-         * 
-         * @type {AdvertisementStore}
-         */
-        this.advertisements = new AdvertisementStore(this);
 
         /**
          * Client's command store
